+++ src/test/regress/sql/numeric.sql
@@ -1,32 +1,32 @@
 ---START---
-CREATE TABLE num_data (gemini_pk serial PRIMARY KEY, id int4, val numeric(210, 10));
+CREATE TABLE num_data (gemini_pk serial PRIMARY KEY, id int8, val numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_add (gemini_pk serial PRIMARY KEY, id1 int4, id2 int4, expected numeric(210, 10));
+CREATE TABLE num_exp_add (gemini_pk serial PRIMARY KEY, id1 int8, id2 int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_sub (gemini_pk serial PRIMARY KEY, id1 int4, id2 int4, expected numeric(210, 10));
+CREATE TABLE num_exp_sub (gemini_pk serial PRIMARY KEY, id1 int8, id2 int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_div (gemini_pk serial PRIMARY KEY, id1 int4, id2 int4, expected numeric(210, 10));
+CREATE TABLE num_exp_div (gemini_pk serial PRIMARY KEY, id1 int8, id2 int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_mul (gemini_pk serial PRIMARY KEY, id1 int4, id2 int4, expected numeric(210, 10));
+CREATE TABLE num_exp_mul (gemini_pk serial PRIMARY KEY, id1 int8, id2 int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_sqrt (gemini_pk serial PRIMARY KEY, id int4, expected numeric(210, 10));
+CREATE TABLE num_exp_sqrt (gemini_pk serial PRIMARY KEY, id int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_ln (gemini_pk serial PRIMARY KEY, id int4, expected numeric(210, 10));
+CREATE TABLE num_exp_ln (gemini_pk serial PRIMARY KEY, id int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_log10 (gemini_pk serial PRIMARY KEY, id int4, expected numeric(210, 10));
+CREATE TABLE num_exp_log10 (gemini_pk serial PRIMARY KEY, id int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_exp_power_10_ln (gemini_pk serial PRIMARY KEY, id int4, expected numeric(210, 10));
+CREATE TABLE num_exp_power_10_ln (gemini_pk serial PRIMARY KEY, id int8, expected numeric(210, 10));
 ---END---
 ---START---
-CREATE TABLE num_result (gemini_pk serial PRIMARY KEY, id1 int4, id2 int4, result numeric(210, 10));
+CREATE TABLE num_result (gemini_pk serial PRIMARY KEY, id1 int8, id2 int8, result numeric(210, 10));
 ---END---
 ---START---
 -- ******************************
